+++ tests/test_hvac_csc.py
@@ -47,11 +47,14 @@ flake8.configure_logging(1)
 
 
 class CscTestCase(salobj.BaseCscTestCase, unittest.IsolatedAsyncioTestCase):
-    def basic_make_csc(self, initial_state, config_dir, simulation_mode, **kwargs):
+    def basic_make_csc(
+        self, initial_state, config_dir, simulation_mode, settings_to_apply, **kwargs
+    ):
         return hvac.HvacCsc(
             initial_state=initial_state,
             config_dir=config_dir,
             simulation_mode=simulation_mode,
+            settings_to_apply=settings_to_apply,
             start_telemetry_publishing=False,
         )
 
@@ -60,6 +63,7 @@ class CscTestCase(salobj.BaseCscTestCase, unittest.IsolatedAsyncioTestCase):
             initial_state=salobj.State.STANDBY,
             config_dir=None,
             simulation_mode=1,
+            settings_to_apply="",
         ):
             await self.check_standard_state_transitions(
                 enabled_commands=(),
@@ -76,7 +80,10 @@ class CscTestCase(salobj.BaseCscTestCase, unittest.IsolatedAsyncioTestCase):
 
     async def test_version(self):
         async with self.make_csc(
-            initial_state=salobj.State.STANDBY, config_dir=None, simulation_mode=1
+            initial_state=salobj.State.STANDBY,
+            config_dir=None,
+            simulation_mode=1,
+            settings_to_apply="",
         ):
             await self.assert_next_sample(
                 self.remote.evt_softwareVersions,
@@ -131,6 +138,7 @@ class CscTestCase(salobj.BaseCscTestCase, unittest.IsolatedAsyncioTestCase):
             initial_state=salobj.State.STANDBY,
             config_dir=None,
             simulation_mode=1,
+            settings_to_apply="",
         ):
             await salobj.set_summary_state(
                 remote=self.remote, state=salobj.State.ENABLED
@@ -222,6 +230,7 @@ class CscTestCase(salobj.BaseCscTestCase, unittest.IsolatedAsyncioTestCase):
             initial_state=salobj.State.STANDBY,
             config_dir=None,
             simulation_mode=1,
+            settings_to_apply="",
         ):
             await salobj.set_summary_state(
                 remote=self.remote, state=salobj.State.ENABLED